MW-377: Speed up retrieving previous requisition
Instead of retrieving previous requisition to get closing balance for the
given product, the tool will save closing balances of created requisitions
and use them when the next requisition will be created.

The tool should retrieve previous requisition only if this is required so
only if there is no data in memory.